requirements.txt
django>=3.2,<4.0
psycopg2-binary>=2.9.9,<3.0
python-dotenv>=1.0.0,<2.0
pytest>=7.4.3,<8.0
pytest-django>=4.7.0,<5.0
gunicorn>=20.0.4
django-cors-headers>=4.3.1 
djangorestframework>=3.14.0
Pillow>=10.2.0
